+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v2.0.0
 
-- n/a
+- use() now only accepts a string as base URL for the request helper functions.
+- remove exitOnFail(), the session always quits upon a test failing.
 
 ## v1.0.0 - v1.0.5